feat: add share_plan/unshare_plan MCP tools
Extracts the course/group snapshot builder used to hydrate registrations into src/lib/plan/build-registration-courses.ts (shared by the client fetcher and the new MCP tools) so an agent can publish a plan's current selection as a public read-only link without going through the browser.
